pyproject.toml
[tool.poetry]
name = "phasegen"
version = "0.0.1-alpha"
description = "Generation and inference on exact moments of coalescent distributions."
authors = ["Sendrowski <sendrowski.janek@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
typing-extensions = "^4.5.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
jsonpickle = "^3.0.0"
multiprocess = "^0.70.12"
pyyaml = "^6.0"
tqdm = "^4.60.0"
scipy = "^1.10.1"

# problems installing msprime with poetry
[tool.poetry.group.dev.dependencies]
urllib3 = "^1.26.0" # problems with poetry and urllib3 version 2
sphinx-rtd-theme = "^1.2.0"
sphinx-autodoc-typehints = "^1.23.0"
sphinx-copybutton = "^0.5.2"
myst-nb = "^0.17.2"
sphinx = "5.3.0"
pygments = "^2.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
